+++ apps/frontend/src/app/pages/transport-problem/table/transport-table.module.ts
@@ -5,10 +5,9 @@ import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatInputModule} from '@angular/material/input';
 import {MatTableModule} from '@angular/material/table';
 
-import {InputTableModule} from '../../../components/input-table/input-table.module';
+import {InputTableModule} from '@components/input-table/input-table.module';
 
 import {TransportTableComponent} from './transport-table.component';
-import {TransportTableService} from './transport-table.service';
 
 @NgModule({
   declarations: [TransportTableComponent],
@@ -20,7 +19,6 @@ import {TransportTableService} from './transport-table.service';
     MatInputModule,
     MatTableModule,
   ],
-  providers: [TransportTableService],
   exports: [TransportTableComponent],
 })
 export class TransportTableModule {}
