+++ apps/frontend/src/app/pages/transport-problem/transport-problem.module.ts
@@ -3,7 +3,7 @@ import {NgModule} from '@angular/core';
 import {MatButtonModule} from '@angular/material/button';
 import {MatSelectModule} from '@angular/material/select';
 
-import {TransportTableModule} from './transport-table/transport-table.module';
+import {TransportTableModule} from './table/transport-table.module';
 import {TransportProblemComponent} from './transport-problem.component';
 import {TransportProblemService} from './transport-problem.service';
