--- apps/frontend/src/app/pages/transport-problem/transport-table/transport-table.template.html
@@ -1,21 +1,0 @@
-<form>
-  <div class="input-table-wrapper">
-    <input-table
-      key="costs"
-      [rows]="storages$ | async"
-      [columns]="shops$ | async"
-      (tableChange)="test($event)"
-    ></input-table>
-    <input-table key="storageStocks" [rows]="storages$ | async"></input-table>
-    <input-table key="shopDemands" [columns]="shops$ | async"></input-table>
-    <button
-      mat-flat-button
-      color="primary"
-      type="button"
-      (click)="clearTable()"
-    >
-      Clear
-    </button>
-  </div>
-  <button mat-flat-button color="primary" type="submit">Calculate</button>
-</form>
