+++ apps/frontend/src/app/shared/helpers/array.helper.ts
@@ -1,0 +1,7 @@
+/** Add numbers in an array and ignores undefined or null values. */
+export function sum(array: Array<number | null | undefined>) {
+  return array.reduce((acc, curr) => {
+    if (!acc || !curr) return curr;
+    return acc + curr;
+  });
+}
