+++ apps/frontend/src/app/shared/types/table.types.ts
@@ -1,0 +1,2 @@
+export type Table = Array<{[key: string]: number | null}>;
+export type RowDefs = Array<string>;
